Add program to display characters on a 2x16 LCD
The program initializes and clears the display at startup. It uses
the X and Y registers to hold the character to display and the column
to display it at, respectively. Only ASCII printable characters
($20-$7e) are displayed, wrapping around once it gets to $7f, though
it starts by displaying 'A' ($41).