Add delay routine for milliseconds
The delay routine makes use of 65C02 extensions, so enable them in the
Makefile with -c02 flag. The assembly has been optimized and adjusted
for a 1.8432 MHz clock instead of the original 1 MHz clock used by Ben
Eater's design to make it easier to run an ACIA chip off the same clock
in the future.

Also add a true "blink" program to test the code and validate the
timing is corre with an oscilloscope.