Add interrupt counting example
This shows how the interrupt request line can be used to sho
Counts how many times the interrupt routine is called and displays it on
the LCD. By hooking a button up to the IRQB line, we can manually
trigger the interrupt routine. We can then play a "game" by seeing
how quickly we can press the button, resulting in a lower count.

https://www.youtube.com/watch?v=DlEa8kd7n3Q