Update address decode logic and register definitions
Now RAM takes bottom 32k of address space, except for the top page,
which is reserved for I/O. Using top 2 address lines of the page (bottom
nibble) allows for 4 I/O devices which all get access to 64 addresses
each. The ATF22V10 has more lines which could be used to further
subdivide the page (I think you could do 16 devices with 16 addresses
each), but so far that is not needed (I am only using 3 of the 4 I/O
devices anyway).

At some pont I should fix the address line pin organization on the PLD
to make it a little more sensible, but it will work for now.