Add start of K65 firmware
Currently, the firmware only does the following:
- initialize the stack pointer to $ff
- initialize the LCD
- displays a message
- loop forever

The following "syscalls" are exported by the firmware:
- lcd_clear
- lcd_home
- lcd_line_two
- print_hex
- delay
- delayms

The list of syscalls and their addresses are generated by a special make
rule which parses the k65 listing for any "*_syscall" labels. These
labels have the syscall suffix removed and then they are listed in
syscalls.inc.s for applications to include. These syscalls are stub
functions which jmp to the actual function address wherever the function
ends up getting assembled at. This addes a 3 cycle delay, but the
benefit is that it allows binary compatibility to be easily maintained.
This will be especially beneficial once SD card loading is enabled,
since otherwise any firmware rebuild would require rebuilding
applications and re-copying them the SD card.