Add SD Commodore-style PRG loading example
Load a Commodore-style PRG file from the first 256 bytes of the SD
card. A PRG file contains a load address header, which can be
automatically created with the -cbm-prg vasm flag. The first two bytes
will be read and the remaining 254 bytes will be read in to the address
specified in the previous two bytes. Unfortunately, the 6502 does not
seem to support an indirect jsr, so a trampoline needs to be used - you
jsr to the trampoline and the trampoline does an indirect jmp to the
loaded program.

In addition, add a hello world variant, which is only designed to be run
from RAM (eg. loaded from the SD card). This works by using subroutines
which are defined by the ROM and must exist in fixed locations in order
for this to work. This PRG can be loaded on to the SD card using dd, eg.
`sudo dd if=hello_world_load.prg of=/dev/sdd`, substituting whatever the
correct path to the actual SD card device (not the first partition!).