Use SPI for SDIO
Use the SPI bus to talk to the SD card. This is much faster than
bit-banging and frees up a lot of GPIO pins on VIA2. For comparison, the
SPI interface currently operates at 444 kHz slow speed and 1 MHz at fast
speed while bit banging was about 66 kHz max.

Since we're not using handshake mode on the VIA1 ports, we can use the
CA2 pin as a GPIO for use as the SD card CS to free up another GPIO pin.
If we had more than one SPI device, we may want to use two GPIO pins and
a selector instead. We use the CA2 pin on VIA1, since it may be
interesting to try using handshaking mode on VIA2 for parallel port
devices.