Load FAT16 program by name
fat16_load_prg now takes in an 11-byte, padded filename in R1 which is
compared with the directory entry filenames to determine which program to load.

In addition, the directory entry loop handling was updated to use
indirect-y loads using y as the offset from the currnet directory entry
pointer, instead of absolute-y loads using y as the offset from the
start of the directory entries and absolute addresses to each field.

This enables all 32 directory entries in the entire 512-byte sector to
be read. There's still no support for loading entries past the first 32,
but that wouldn't be too difficult to do from dir_loop_next if the code
to load a directory entry was factored out in to a separate function.

The loading message was also moved outside the body of the loop. This
makes the loop body tighter and more clear.