Fix typos and improve FAT16 diagram
The diagram claimed that there was the following FAT chain:

11 -> 15 -> 13 -> 14 -> 16 -> 17

However, FAT entry 11 was 0 (unused), while 19 pointed at 15. In
addition, FAT entry 17 pointed to FAT entry 18, but the chain did not
show this.

For improvements, make the second chain a bit more complicated by skipping an
cluster. Also, list the free clusters for completeness.